requirements.txt
Django>=4.0.5
djangorestframework>=3.13.1
psycopg2-binary>=2.9.3
gunicorn>=20.1.0
scikit-learn>=1.1.2
numpy>=1.21.5
pandas>=1.4.3